#1 종목 정보 수집 및 알림 기능 변경

+++ module-analyzer/src/main/java/analyzer/global/config/RedisConfig.java
@@ -24,8 +24,8 @@ public class RedisConfig {
     }
 
     @Bean
-    public RedisTemplate<String, Integer> redisTemplate() {
-        RedisTemplate<String, Integer> template = new RedisTemplate<>();
+    public RedisTemplate<String, Double> redisTemplate() {
+        RedisTemplate<String, Double> template = new RedisTemplate<>();
         template.setConnectionFactory(redisConnectionFactory());
         template.setKeySerializer(new StringRedisSerializer());
         template.setHashKeySerializer(new StringRedisSerializer());
